split DTreeGraphConnectivity into multiple smaller classes
