--- src/main/java/com/powsybl/openloadflow/graph/DTreeGraphConnectivity.java
@@ -1,1054 +1,0 @@
-/**
- * Copyright (c) 2026, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- * SPDX-License-Identifier: MPL-2.0
- */
-package com.powsybl.openloadflow.graph;
-
-import org.apache.commons.lang3.tuple.ImmutablePair;
-import org.apache.commons.lang3.tuple.Pair;
-
-import java.util.*;
-
-/**
- * D-Tree implementation from <cite>Qing Chen, Oded Lachish, Sven Helmer, and Michael H. Böhlen. Dynamic
- * Spanning Trees for Connectivity Queries on Fully-dynamic Undirected
- * Graphs. PVLDB, 15(11): 3263 - 3276, 2022.
- * doi:10.14778/3551793.3551868</cite>. An extended version is available at
- * <a href="https://arxiv.org/pdf/2207.06887">https://arxiv.org/pdf/2207.06887</a>
- *
- * <p>
- * This implementation also includes one optimization (see {@link DTGraph#insertNonTreeEdge(DTGraph.DTNode, DTGraph.DTNode, int, DTGraph.DTNode, int, DTGraph.Edge)})
- * from another paper: <cite>Lantian Xu, Dong Wen, Lu Qin, Ronghua Li, Ying Zhang, and Xuemin Lin. 2024.
- * Constant-time Connectivity Querying in Dynamic Graphs. Proc. ACM Manag. Data 2, 6 (SIGMOD), Article 230
- * (December 2024), 23 pages. <a href="https://doi.org/10.1145/3698805">https://doi.org/10.1145/3698805</a></cite>.
- * </p>
- *
- * @author Valentin Carrez {@literal <valentin.carrez at rte-france.com>}
- */
-public class DTreeGraphConnectivity<V, E> extends AbstractGraphConnectivity<V, E, DTreeGraphConnectivity.DTGraph<V, E>> {
-
-    public DTreeGraphConnectivity() {
-        super(new DTGraph<>());
-    }
-
-    @Override
-    protected void updateConnectivity(EdgeRemove<V, E> edgeRemove) {
-        // only invalidate components.
-        // update is done directly in DTGraph because the graph is stored inside the spanning forest.
-        componentSets = null;
-    }
-
-    @Override
-    protected void updateConnectivity(EdgeAdd<V, E> edgeAdd) {
-        // only invalidate components.
-        // update is done directly in DTGraph because the graph is stored inside the spanning forest.
-        componentSets = null;
-    }
-
-    @Override
-    protected void updateConnectivity(VertexAdd<V, E> vertexAdd) {
-        // only invalidate components.
-        // update is done directly in DTGraph because the graph is stored inside the spanning forest.
-        componentSets = null;
-    }
-
-    @Override
-    protected void resetConnectivity(Deque<GraphModification<V, E>> m) {
-        // only invalidate components.
-        // update is done directly in undoTemporaryChanges because the graph is stored inside the spanning forest.
-        componentSets = null;
-    }
-
-    @Override
-    protected void updateComponents() {
-        if (componentSets != null) {
-            return;
-        }
-
-        DTGraph<V, E> graph = getGraph();
-        List<DTGraph<V, E>.DTNode> roots = graph.roots;
-
-        // sorting roots will sort components as components is a wrapper around roots
-        roots.sort(Comparator.comparingInt((DTGraph<V, E>.DTNode root) -> root.size).reversed());
-        for (int i = 0; i < graph.roots.size(); i++) {
-            roots.get(i).rootIndex = i;
-        }
-
-        componentSets = graph.components;
-    }
-
-    @Override
-    protected int getQuickComponentNumber(V vertex) {
-        return getGraph().rootOf(vertex).rootIndex;
-    }
-
-    @Override
-    public int getNbConnectedComponents() {
-        checkSavedContext();
-        return getGraph().roots.size();
-    }
-
-    @Override
-    public Set<V> getConnectedComponent(V vertex) {
-        checkSavedContext();
-        checkVertex(vertex);
-        return getGraph().vertexToTreeNode.get(vertex).componentView();
-    }
-
-    @Override
-    protected Set<V> getNonConnectedVertices(V vertex) {
-        checkSavedContext();
-        checkVertex(vertex);
-
-        DTGraph<V, E> graph = getGraph();
-        DTGraph<V, E>.DTNode excludedTree = graph.rootOf(vertex);
-
-        Set<V> components = new HashSet<>();
-        for (DTGraph<V, E>.DTNode root : graph.roots) {
-            if (root != excludedTree) {
-                components.addAll(root.componentView());
-            }
-        }
-
-        return components;
-    }
-
-    @Override
-    protected Set<V> getVerticesNotInMainComponent(V mainComponentVertex) {
-        // first determine the excluded tree: either the tree containing
-        // the mainComponentVertex, either the biggest tree
-        DTGraph<V, E>.DTNode excludedTree = getMainComponentRoot(mainComponentVertex);
-        return new VerticesNotInMainComponent(excludedTree);
-    }
-
-    private final class VerticesNotInMainComponent extends AbstractSetView<V> {
-        private final DTGraph<V, E>.DTNode excludedTree;
-        private int size = -1;
-
-        VerticesNotInMainComponent(DTGraph<V, E>.DTNode excludedTree) {
-            this.excludedTree = excludedTree;
-        }
-
-        @Override
-        public Iterator<V> iterator() {
-            return new VerticesNotInMainComponentIterator(excludedTree.findRoot());
-        }
-
-        @Override
-        public boolean contains(Object o) {
-            if (o != null) {
-                return getGraph().rootOf((V) o) != excludedTree.findRoot();
-            }
-
-            return false;
-        }
-
-        @Override
-        public int size() {
-            if (size < 0) {
-                size = 0;
-
-                DTGraph<V, E>.DTNode excludedTreeRoot = excludedTree.findRoot();
-                size = getGraph().roots.stream()
-                        .filter(root -> root != excludedTreeRoot)
-                        .mapToInt(root -> root.size)
-                        .sum();
-            }
-
-            return size;
-        }
-    }
-
-    private class VerticesNotInMainComponentIterator implements Iterator<V> {
-
-        private final DTGraph<V, E>.DTNode excludedTree;
-        private int index = 0;
-        private Iterator<V> curIt;
-
-        VerticesNotInMainComponentIterator(DTGraph<V, E>.DTNode excludedTree) {
-            this.excludedTree = excludedTree;
-        }
-
-        @Override
-        public boolean hasNext() {
-            if (curIt != null && curIt.hasNext()) {
-                return true;
-            }
-
-            DTGraph<V, E> graph = getGraph();
-            while (index < graph.roots.size()) {
-                DTGraph<V, E>.DTNode next = graph.roots.get(index);
-                index++;
-
-                if (next != excludedTree) {
-                    curIt = getGraph().iterator(next);
-                    return true;
-                }
-            }
-
-            return false;
-        }
-
-        @Override
-        public V next() {
-            if (!hasNext()) {
-                throw new NoSuchElementException();
-            }
-
-            return curIt.next();
-        }
-    }
-
-    /**
-     * @param mainComponentVertex a vertex in the main component tree, may be null
-     * @return the root of the tree containing mainComponentVertex, if not null,
-     * or the root of the biggest tree
-     */
-    private DTGraph<V, E>.DTNode getMainComponentRoot(V mainComponentVertex) {
-        DTGraph<V, E> graph = getGraph();
-
-        if (mainComponentVertex != null) {
-            return graph.rootOf(mainComponentVertex);
-        } else {
-            DTGraph<V, E>.DTNode biggestRoot = graph.roots.getFirst();
-
-            for (int i = 1; i < graph.roots.size(); i++) {
-                DTGraph<V, E>.DTNode root = graph.roots.get(i);
-                if (root.size > biggestRoot.size) {
-                    biggestRoot = root;
-                }
-            }
-
-            return biggestRoot;
-        }
-    }
-
-    @Override
-    public boolean supportTemporaryChangesNesting() {
-        return true;
-    }
-
-    public long computeSd() {
-        return getGraph().sumOfDistances();
-    }
-
-    public static final class DTGraph<V, E> implements GraphModel<V, E> {
-
-        /** map a vertex to a node in a spanning tree */
-        private final Map<V, DTNode> vertexToTreeNode = new HashMap<>();
-        /** map an edge to an edge in a spanning tree */
-        private final Map<E, Edge> edges = new HashMap<>();
-
-        /** the list of tree roots. Roots are maintained in a way such that
-         * the value of the attribute 'rootIndex' of the DTNode at index i is i.
-         * In other words: roots.get(i).rootIndex == i */
-        private final List<DTNode> roots = new ArrayList<>();
-
-        private final AllComponentsView components = new AllComponentsView();
-
-        public long sumOfDistances() {
-            long sum = 0;
-
-            for (DTNode node : vertexToTreeNode.values()) {
-                sum += node.findRootWithDepth().getValue();
-            }
-
-            return sum;
-        }
-
-        /**
-         * Return the root of the tree in which {@code vertex} is and
-         * eventually reroot the tree if a new centroid is found.
-         *
-         * @param vertex the vertex whose tree root is to be returned.
-         * @return the root of the tree in which {@code vertex} is.
-         */
-        DTNode rootOf(V vertex) {
-            return vertexToTreeNode.get(vertex).findRootOptReroot();
-        }
-
-        @Override
-        public void addEdge(V u, V v, E e) {
-            if (containsEdge(e)) {
-                return;
-            }
-
-            DTNode nodeU = getNodeThrowIfInexistent(u);
-            DTNode nodeV = getNodeThrowIfInexistent(v);
-
-            // update edges
-            Edge edge = new Edge(nodeU, nodeV, e, false);
-            edges.put(e, edge);
-
-            // update spanning trees
-            Pair<DTNode, Integer> rootUdist = nodeU.findRootWithDepth();
-            Pair<DTNode, Integer> rootVdist = nodeV.findRootWithDepth();
-
-            if (rootUdist.getKey() == rootVdist.getKey()) {
-                // insert non tree edge
-                edge.treeEdge = insertNonTreeEdge(rootUdist.getKey(), nodeU, rootUdist.getValue(), nodeV, rootVdist.getValue(), edge);
-            } else {
-                // insert tree edge
-                edge.treeEdge = true;
-                insertTreeEdge(rootUdist.getKey(), nodeU, rootVdist.getKey(), nodeV, edge);
-            }
-        }
-
-        private DTNode getNodeThrowIfInexistent(V v) {
-            DTNode node = vertexToTreeNode.get(v);
-            if (node == null) {
-                throw new IllegalArgumentException("no such vertex in graph: " + v);
-            }
-
-            return node;
-        }
-
-        /**
-         * Insert a non tree edge between {@code nodeU} (whose depth is {@code depthU})
-         * and {@code nodeV} (whose depth is {@code depthV}). The two nodes must be in the
-         * same tree rooted at {@code root}.
-         * <p>
-         * If the difference of depth, delta, is less than two, the edge is inserted
-         * as a non-tree edge. Otherwise, assuming depthU < depthV, the delta / 2 - 1 ancestor of
-         * {@code nodeU} is unlinked from the tree. Then {@code nodeU} and {@code nodeV} are
-         * linked with a tree edge. In this case, the inserted edge is in fact a tree edge
-         * and the method return {@code true}
-         * </p>
-         *
-         * <p>
-         * The original DTree paper uses delta - 2 instead of delta / 2 - 1. But a more recent
-         * article indicates better results with delta / 2 - 1. Experimentation confirms this,
-         * the average depth is smaller with the new upper bound.
-         * </p>
-         *
-         * @param root the root of the tree in which an edge is to be added.
-         * @param nodeU one endpoint of the edge to add.
-         * @param depthU the depth of {@code nodeU}.
-         * @param nodeV the other endpoint of the edge to add.
-         * @param depthV the depth of {@code nodeU}
-         * @param edge edge linking {@code nodeU} and {@code nodeV}
-         * @return {@code true} if the edge inserted is a tree edge. This is true if |depthU - depthV| >= 2.
-         */
-        private boolean insertNonTreeEdge(DTNode root, DTNode nodeU, int depthU, DTNode nodeV, int depthV, Edge edge) {
-            DTNode deep;
-            DTNode shallow;
-            int delta;
-
-            if (depthU <= depthV) {
-                shallow = nodeU;
-                deep = nodeV;
-                delta = depthV - depthU;
-            } else {
-                shallow = nodeV;
-                deep = nodeU;
-                delta = depthU - depthV;
-            }
-
-            if (delta < 2) {
-                // no changes in the BFS tree
-                nodeU.nonTreeEdges.add(edge);
-                nodeV.nonTreeEdges.add(edge);
-                return false;
-            } else {
-                // get the (delta / 2 - 1) DTNode.
-                DTNode ancestor = deep;
-                for (int j = 0; j < delta / 2 - 1; j++) {
-                    ancestor = ancestor.parent;
-                }
-
-                // replace the edge between ancestor and its parent by a non tree edge
-                ancestor.parent.nonTreeEdges.add(ancestor.parentEdge);
-                ancestor.nonTreeEdges.add(ancestor.parentEdge);
-                ancestor.parentEdge.treeEdge = false;
-                ancestor.unlink();
-
-                // updating roots is useless because 'deep' will be
-                // connected to 'shallow' juste after. Updating is also impossible
-                // because the tree created by the previous unlink isn't in 'roots'
-                deep.makeRoot(false);
-                deep.link(root, shallow, edge);
-                return true;
-            }
-        }
-
-        /**
-         * Insert a tree edge between {@code nodeU} (in tree rooted at {@code rootU})
-         * and {@code nodeV} (in tree rooted at {@code rootV}).
-         * Assuming the size of rootU is less than the size of rootV, we simply
-         * make {@code nodeU} a root and link it with {@code nodeV}.
-         *
-         * @param rootU {@code nodeU} tree root
-         * @param nodeU one endpoint of the edge to add.
-         * @param rootV {@code nodeV} tree root
-         * @param nodeV the other endpoint of the edge to add.
-         * @param edge edge linking {@code nodeU} and {@code nodeV}
-         */
-        private void insertTreeEdge(DTNode rootU, DTNode nodeU, DTNode rootV, DTNode nodeV, Edge edge) {
-            if (rootU.size < rootV.size) {
-                nodeU.makeRoot(true);
-                nodeU.link(rootV, nodeV, edge);
-                removeRoot(nodeU);
-            } else {
-                nodeV.makeRoot(true);
-                nodeV.link(rootU, nodeU, edge);
-                removeRoot(nodeV);
-            }
-        }
-
-        @Override
-        public void removeEdge(E e) {
-            Edge edge = edges.remove(e);
-            if (edge == null) {
-                return;
-            }
-
-            if (edge.treeEdge) {
-                removeTreeEdge(edge.nodeU, edge.nodeV);
-            } else {
-                removeNonTreeEdge(edge);
-            }
-        }
-
-        /**
-         * Remove the tree edge between {@code nodeU} and {@code nodeV}.
-         * Assuming nodeU is a child of nodeV, this is a two steps process :
-         * <ol>
-         *     <li>Unlink nodeU from nodeV. This creates two trees with a smaller one called {@code small},</li>
-         *     <li>Search for a replacement edge and a potential new centroid by iterating over {@code small}.</li>
-         *     <ul>
-         *         <li>if one is found, it is a non-tree edge so it is removed and then added as a tree edge</li>
-         *         <li>if none is found, fix the centroid property</li>
-         *     </ul>
-         * </ol>
-         *
-         * @param nodeU one endpoint of the edge to remove.
-         * @param nodeV the other endpoint of the edge to remove.
-         */
-        private void removeTreeEdge(DTNode nodeU, DTNode nodeV) {
-            DTNode child;
-
-            if (nodeU == nodeV.parent) {
-                child = nodeV;
-            } else {
-                child = nodeU;
-            }
-
-            // unlink child from its parent
-            DTNode otherTree = child.unlink();
-            addRoot(child);
-
-            DTNode small;
-            if (child.size < otherTree.size) {
-                small = child;
-            } else {
-                small = otherTree;
-            }
-
-            // try to reconnect them
-            replace(small);
-        }
-
-        private void replace(DTNode rootSmall) {
-            DTNode newRoot = null; // a potential new root in case no replacement edge is found
-
-            // iterate over the nodes of rootSmall using a BFS.
-            ArrayDeque<DTNode> queue = new ArrayDeque<>();
-            queue.offer(rootSmall);
-
-            while (!queue.isEmpty()) {
-                DTNode n = queue.poll();
-
-                // search for a new centroid
-                if (n != rootSmall && n.size > rootSmall.size / 2) {
-                    newRoot = n;
-                }
-
-                // search for a replacement edge
-                for (Edge nonTreeEdge : n.nonTreeEdges) {
-                    DTNode oppNode = nonTreeEdge.opposite(n);
-                    DTNode oppRoot = oppNode.findRoot();
-
-                    if (oppRoot != rootSmall) {
-                        // found a replacement edge
-                        removeNonTreeEdge(nonTreeEdge);
-                        insertTreeEdge(rootSmall, n, oppRoot, oppNode, nonTreeEdge);
-                        nonTreeEdge.treeEdge = true;
-
-                        return;
-                    }
-                }
-
-                // add all children to the queue
-                DTNode child = n.firstChild;
-                while (child != null) {
-                    queue.add(child);
-                    child = child.nextSibling;
-                }
-            }
-
-            // fix centroid property
-            if (newRoot != null) {
-                newRoot.makeRoot(true);
-            }
-        }
-
-        /**
-         * Remove a non tree edge between {@code edge.nodeU} and {@code edge.nodeV}.
-         * @param edge the edge to remove.
-         */
-        private void removeNonTreeEdge(Edge edge) {
-            edge.nodeU.nonTreeEdges.remove(edge);
-            edge.nodeV.nonTreeEdges.remove(edge);
-        }
-
-        private void addRoot(DTNode node) {
-            node.rootIndex = roots.size();
-            roots.add(node);
-        }
-
-        private void removeRoot(DTNode node) {
-            // update roots, swapping 'node' and the last element of roots
-            DTNode last = roots.removeLast();
-            if (node != last) {
-                last.rootIndex = node.rootIndex;
-                roots.set(last.rootIndex, last);
-            }
-        }
-
-        private Iterator<V> iterator(DTNode root) {
-            return new DFSIterator(root);
-        }
-
-        @Override
-        public void addVertex(V v) {
-            if (containsVertex(v)) {
-                return;
-            }
-
-            DTNode newNode = new DTNode(v);
-            vertexToTreeNode.put(v, newNode);
-            addRoot(newNode);
-        }
-
-        @Override
-        public void removeVertex(V v) {
-            if (!containsVertex(v)) {
-                return;
-            }
-
-            for (E edge : getNeighborEdgesOf(v)) {
-                removeEdge(edge);
-            }
-            DTNode root = vertexToTreeNode.remove(v);
-            removeRoot(root);
-        }
-
-        @Override
-        public boolean containsVertex(V vertex) {
-            return vertexToTreeNode.containsKey(vertex);
-        }
-
-        @Override
-        public boolean containsEdge(E edge) {
-            return edges.containsKey(edge);
-        }
-
-        @Override
-        public V getEdgeSource(E edge) {
-            return switch (edges.get(edge)) {
-                case null -> null;
-                case Edge e -> e.nodeU.vertex;
-            };
-        }
-
-        @Override
-        public V getEdgeTarget(E edge) {
-            return switch (edges.get(edge)) {
-                case null -> null;
-                case Edge e -> e.nodeV.vertex;
-            };
-        }
-
-        @Override
-        public Set<E> getEdgesBetween(V vertex1, V vertex2) {
-            throw new UnsupportedOperationException();
-        }
-
-        @Override
-        public Set<E> getEdges() {
-            throw new UnsupportedOperationException();
-        }
-
-        @Override
-        public Set<E> getNeighborEdgesOf(V v) {
-            return vertexToTreeNode.get(v).getNeighborEdges();
-        }
-
-        @Override
-        public int getNeighborEdgeCountOf(V v) {
-            throw new UnsupportedOperationException();
-        }
-
-        @Override
-        public Set<V> getVertices() {
-            return vertexToTreeNode.keySet();
-        }
-
-        @Override
-        public List<V> getNeighborVerticesOf(V v) {
-            throw new UnsupportedOperationException();
-        }
-
-        /**
-         * A DTNode (Dynamic Tree Node) is a node in a spanning tree.
-         * Each DTNode maintains the following information:
-         * <ul>
-         *     <li>the vertex in the graph,</li>
-         *     <li>the size of the subtree,</li>
-         *     <li>its parent in the tree and the edge linking them,</li>
-         *     <li>its children in the tree and the edges linking them,</li>
-         *     <li>all non tree edges having at least one endpoint that is the DTNode</li>
-         *     <li>if the node is a root, its index in the list of {@link DTGraph#roots}</li>
-         * </ul>
-         *
-         * <p>
-         * However, children are stored in a particular way, allowing
-         * fast iteration over a tree and insertion and removal of a child,
-         * but slow access to an arbitrary element. Instead of storing them
-         * in a list or a map, each DTNode has a pointer to its previous sibling,
-         * its next sibling and its first child. In other words, the children of
-         * a DTNode are stored in a doubly-linked list. A DTNode stores the first
-         * element in this list and is also used as an element in its parent doubly
-         * linked list of children.
-         * </p>
-         *
-         * <p>
-         * Example:
-         * <pre>
-         * +----- first child ------ 1
-         * |                         ^
-         * |                         |
-         * |                      parent
-         * |                         |
-         * | +-----------------------+-----------------------+
-         * v/                        |                        \
-         * 2 <-- previous sibling -- 3 <-- previous sibling -- 4
-         *  \_____ next sibling _____^\_____ next sibling _____^
-         * </pre>
-         * X --> Y indicates that X contains a pointer to Y.
-         *</p>
-         *
-         * <p>
-         * This complex structure allows fast insertion and removal as we only need
-         * to update the sibling list and eventually the first child pointer. But the
-         * biggest advantage is that it allows fast iteration of a tree with 0 memory
-         * allocations by only following pointers. See {@link DFSIterator}
-         * </p>
-         *
-         */
-        private final class DTNode {
-
-            private final V vertex;
-
-            // the size of this subtree
-            private int size;
-
-            private DTNode parent = null;
-            private Edge parentEdge = null;
-
-            // the children of this node. They are stored in a doubly linked list
-            // firstChild is the head of the linked list. previousSibling and nextSibling
-            // are used to navigate the list.
-            private DTNode firstChild = null;
-            private DTNode previousSibling = null;
-            private DTNode nextSibling = null;
-
-            private final Set<Edge> nonTreeEdges = new HashSet<>();
-
-            // index in the list of roots, valid only if this node is a root
-            private int rootIndex;
-
-            private ComponentView componentView = null;
-
-            DTNode(V vertex) {
-                this.vertex = vertex;
-                this.size = 1;
-            }
-
-            /**
-             * Make this DTNode the root of the tree in which it is. This is a two steps process
-             * with an intermediate optional operation:
-             * <ol>
-             *     <li>Swap parent-child relationship for each DTNode from this dTNode to the original root</li>
-             *     <li>Optionally, update the list of roots. For most cases, it must be {@code true}</li>
-             *     <li>Update the subtree size attribute from the original root to the new root (this DTNode)</li>
-             * </ol>
-             *
-             * @param updateRoots {@code true} to update the list of roots
-             */
-            private void makeRoot(boolean updateRoots) {
-                if (parent == null) {
-                    return;
-                }
-
-                DTNode child = this;
-                DTNode oldParent = child.parent;
-                Edge oldParentEdge = child.parentEdge;
-                oldParent.removeChildUnchecked(child); // remove before making parentEdge null
-
-                this.parent = null;
-                this.parentEdge = null;
-
-                // swap parent/child relation
-                while (oldParent != null) {
-                    DTNode greatParent = oldParent.parent;
-                    Edge greatParentEdge = oldParent.parentEdge;
-
-                    // At this point:
-                    // - 'oldParent' is in the linked list of children of 'greatParent', and must be
-                    //   removed from it because adding 'oldParent' as a child of 'child' will break
-                    //   this linked list.
-                    // - 'child' is NOT in the linked list of children of 'oldParent'.
-                    //   It was removed by the last iteration or before entering in the loop (for the first iteration)
-                    // - the parent of 'oldParent' aka 'greatParent' should be changed to child
-                    if (greatParent != null) {
-                        greatParent.removeChildUnchecked(oldParent);
-                    }
-
-                    child.addChildUnchecked(oldParent);
-                    oldParent.parent = child;
-                    oldParent.parentEdge = oldParentEdge;
-
-                    // At this point:
-                    // - 'oldParent' isn't anymore is the linked list of child of 'greatParent'
-                    // - 'oldParent' is a child of 'child'
-
-                    // process to the next parent/child
-                    child = oldParent;
-                    oldParent = greatParent;
-                    oldParentEdge = greatParentEdge;
-                }
-
-                // child is the old root
-                DTNode oldRoot = child;
-
-                // update the list of roots
-                if (updateRoots) {
-                    rootIndex = oldRoot.rootIndex;
-                    roots.set(rootIndex, DTNode.this);
-                }
-
-                // update size attributes, going from oldRoot to this DTNode
-                while (oldRoot.parent != null) {
-                    oldRoot.size -= oldRoot.parent.size;
-                    oldRoot.parent.size += oldRoot.size;
-                    oldRoot = oldRoot.parent;
-                }
-            }
-
-            /**
-             * Make this node a child of {@code parent} whose tree is rooted at {@code parentRoot}.
-             * @param parentRoot root of parent
-             * @param parent node that will become the parent of {@code this}.
-             * @param edge the edge linking {@code this} and {@code parent}
-             */
-            private void link(DTNode parentRoot, DTNode parent, Edge edge) {
-                // first: update parent/child relations
-                parent.addChildUnchecked(this);
-                this.parent = parent;
-                this.parentEdge = edge;
-
-                // next: update size attributes in the parent tree
-                DTNode newCentroid = null;
-                DTNode cur = parent;
-
-                while (cur != null) {
-                    cur.size += this.size;
-
-                    if (newCentroid == null && cur != parentRoot && cur.size > (parentRoot.size + this.size) / 2) {
-                        // the new root is the first node in the path from parent to parentRoot
-                        // such that it contains more than half of the nodes in the merged
-                        // tree. This reduces the sum of distances.
-                        newCentroid = cur;
-                    }
-
-                    cur = cur.parent;
-                }
-
-                // eventually, change the root to a better one
-                if (newCentroid != null) {
-                    newCentroid.makeRoot(true);
-                }
-            }
-
-            /**
-             * Unlink this node from its parent, creating two trees, one
-             * whose root is {@code this} and one whose root is returned and
-             * was previously the root of the linked tree.
-             *
-             * @return the root of the other tree
-             */
-            private DTNode unlink() {
-                Objects.requireNonNull(parent);
-
-                // first step: update size attribute in the parent tree
-                DTNode newTree = this;
-                while (newTree.parent != null) {
-                    newTree = newTree.parent;
-                    newTree.size -= size;
-                }
-
-                // second step: update parent/child relations
-                parent.removeChildUnchecked(this);
-                parent = null;
-                parentEdge = null;
-                return newTree;
-            }
-
-            /**
-             * Add child in the doubly linked list of children.
-             * The child mustn't be in a linked list. No verification
-             * is performed.
-             *
-             * @param child the child node to add
-             */
-            private void addChildUnchecked(DTNode child) {
-                DTNode oldFirstChild = this.firstChild;
-
-                firstChild = child;
-                child.nextSibling = oldFirstChild;
-
-                if (oldFirstChild != null) {
-                    oldFirstChild.previousSibling = child;
-                }
-            }
-
-            /**
-             * Remove child from the doubly linked list of children.
-             * The child must be in the linked list. No verification
-             * is performed.
-             *
-             * @param child the child node to remove
-             */
-            private void removeChildUnchecked(DTNode child) {
-                DTNode prev = child.previousSibling;
-                DTNode next = child.nextSibling;
-
-                child.previousSibling = null;
-                child.nextSibling = null;
-
-                if (prev != null) {
-                    prev.nextSibling = next;
-                } else {
-                    // if there is no 'prev' node, that means
-                    // that 'child' was the first child, and
-                    // we need to update it
-                    firstChild = next;
-                }
-                if (next != null) {
-                    next.previousSibling = prev;
-                }
-            }
-
-            private DTNode findRoot() {
-                DTNode node = this;
-
-                while (node.parent != null) {
-                    node = node.parent;
-                }
-
-                return node;
-            }
-
-            private Pair<DTNode, Integer> findRootWithDepth() {
-                DTNode node = this;
-                int depth = 0;
-
-                while (node.parent != null) {
-                    node = node.parent;
-                    depth++;
-                }
-
-                return new ImmutablePair<>(node, depth);
-            }
-
-            /**
-             * Returns the root of the tree containing this node
-             * and restores the centroid property by rerooting the
-             * tree. This helps reduce the height of the tree.
-             * <strong>Warning</strong>: rerooting the tree may
-             * break connectivity queries.
-             *
-             * @return the root of the tree
-             */
-            private DTNode findRootOptReroot() {
-                DTNode nodeRoot = this;
-                DTNode nodeRootChild = null; // the child of nodeRoot in the path from nodeRoot to node
-
-                while (nodeRoot.parent != null) {
-                    nodeRootChild = nodeRoot;
-                    nodeRoot = nodeRoot.parent;
-                }
-
-                if (nodeRootChild != null && nodeRootChild.size > nodeRoot.size / 2) {
-                    nodeRootChild.makeRoot(true);
-                    nodeRoot = nodeRootChild;
-                }
-
-                return nodeRoot;
-            }
-
-            public Set<V> componentView() {
-                if (componentView == null) {
-                    componentView = new ComponentView(this);
-                }
-
-                return componentView;
-            }
-
-            public Set<E> getNeighborEdges() {
-                Set<E> neighbor = new HashSet<>();
-
-                if (parentEdge != null) {
-                    neighbor.add(parentEdge.edgeData);
-                }
-
-                for (Edge nte : nonTreeEdges) {
-                    neighbor.add(nte.edgeData);
-                }
-
-                DTNode child = firstChild;
-                while (child != null) {
-                    neighbor.add(child.parentEdge.edgeData);
-                    child = child.nextSibling;
-                }
-
-                return neighbor;
-            }
-        }
-
-        private final class AllComponentsView extends AbstractList<Set<V>> {
-
-            @Override
-            public Set<V> get(int index) {
-                return roots.get(index).componentView();
-            }
-
-            @Override
-            public int size() {
-                return roots.size();
-            }
-        }
-
-        private final class ComponentView extends AbstractSetView<V> {
-
-            private final DTNode node;
-
-            ComponentView(DTNode node) {
-                this.node = node;
-            }
-
-            @Override
-            public Iterator<V> iterator() {
-                return new DFSIterator(node.findRoot());
-            }
-
-            @Override
-            public boolean contains(Object o) {
-                if (o != null) {
-                    // node might not be the root anymore, so need to use findRoot on node.
-                    // However, don't use findRootOptReroot on node, it might change the root
-                    // after we got the root of 'o'.
-
-                    return rootOf((V) o) == node.findRoot();
-                }
-
-                return false;
-            }
-
-            @Override
-            public int size() {
-                return node.findRoot().size;
-            }
-        }
-
-        private final class DFSIterator implements Iterator<V> {
-
-            private DTNode cursor;
-
-            /**
-             * Creates a new depth-first iterator starting at the specified root node
-             * and returning node according to the pre-order.
-             *
-             * @param root the root of the tree to traverse. It must be a root otherwise,
-             *             the iterator may visit nodes outside the subtree
-             */
-            DFSIterator(DTNode root) {
-                cursor = root;
-            }
-
-            @Override
-            public boolean hasNext() {
-                return cursor != null;
-            }
-
-            @Override
-            public V next() {
-                if (!hasNext()) {
-                    throw new NoSuchElementException();
-                }
-
-                DTNode next = cursor;
-
-                // Advances to the next node for the next iteration.
-                // The iterator try to:
-                // - descend one level whenever possible,
-                // - otherwise, moves to the next sibling if any,
-                // - otherwise, moves up until it finds a node with
-                //   a next sibling (unvisited by construction) or
-                //   the tree is fully visited.
-
-                if (cursor.firstChild != null) {
-                    cursor = cursor.firstChild;
-                } else if (cursor.nextSibling != null) {
-                    cursor = cursor.nextSibling;
-                } else {
-                    while (cursor != null && cursor.nextSibling == null) {
-                        cursor = cursor.parent;
-                    }
-
-                    if (cursor != null) {
-                        cursor = cursor.nextSibling;
-                    }
-                }
-
-                return next.vertex;
-            }
-        }
-
-        private final class Edge {
-            private final DTNode nodeU;
-            private final DTNode nodeV;
-            private final E edgeData;
-            private boolean treeEdge;
-
-            private Edge(DTNode nodeU, DTNode nodeV, E edgeData, boolean treeEdge) {
-                this.nodeU = nodeU;
-                this.nodeV = nodeV;
-                this.edgeData = edgeData;
-                this.treeEdge = treeEdge;
-            }
-
-            public DTNode opposite(DTNode node) {
-                if (nodeU == node) {
-                    return nodeV;
-                } else {
-                    return nodeU;
-                }
-            }
-        }
-    }
-}
