+++ src/main/java/com/powsybl/openloadflow/graph/dtree/ComponentView.java
@@ -1,0 +1,42 @@
+/**
+ * Copyright (c) 2026, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ * SPDX-License-Identifier: MPL-2.0
+ */
+package com.powsybl.openloadflow.graph.dtree;
+
+import java.util.Iterator;
+
+public class ComponentView<V, E> extends AbstractSetView<V> {
+
+    private final DTNode<V, E> node;
+
+    ComponentView(DTNode<V, E> node) {
+        this.node = node;
+    }
+
+    @Override
+    public Iterator<V> iterator() {
+        return new DFSIterator<>(node.findRoot());
+    }
+
+    @Override
+    public boolean contains(Object o) {
+        if (o != null) {
+            // node might not be the root anymore, so need to use findRoot on node.
+            // However, don't use findRootOptReroot on node, it might change the root
+            // after we got the root of 'o'.
+
+            return node.graph.rootOf((V) o) == node.findRoot();
+        }
+
+        return false;
+    }
+
+    @Override
+    public int size() {
+        return node.findRoot().size;
+    }
+}
