+++ src/main/java/com/powsybl/openloadflow/graph/dtree/DTreeGraphConnectivityFactory.java
@@ -5,7 +5,10 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  * SPDX-License-Identifier: MPL-2.0
  */
-package com.powsybl.openloadflow.graph;
+package com.powsybl.openloadflow.graph.dtree;
+
+import com.powsybl.openloadflow.graph.GraphConnectivity;
+import com.powsybl.openloadflow.graph.GraphConnectivityFactory;
 
 /**
  * @author Valentin Carrez {@literal <valentin.carrez at rte-france.com>}
